Move collect_data_for_theme() and related constants to themes.py
